to_do/backend_todo/requirements.txt
flask==3.0.3
flask-cors==5.0.0
psycopg2-binary==2.9.9 --only-binary=psycopg2-binary